Comment before action_on_unpermitted_parameters
The `config/environments/development.rb` file has comments before each
of its settings, but a comment is missing when we insert
`action_on_unpermitted_parameters`. The lack of symmetry made me sad,
and the fact that we may hand the project off to another dev team makes
me think we should document it.

The `#configure_environment` method was inserting spaces for us, but
those spaces were throwing the formatting off for multi-line insertions.
Fall back to `#inject_into_file` instead.