Do not ignore errors when checking for foreman
The `&>` syntax is not valid sh. Instead of changing it to valid sh
syntax (`2>&1 > /dev/null`), we can instead drop the `&` entirely.

If `command` prints to stderr, `foreman` may or may not be installed. In
this case the user will want to know that there was an error; telling
them that `foreman` is not installed masks the error, making debugging
trickier.