Turn off Travis email notifications
The documentation gives this as the correct way to turn off build notifications:

  notifications:
    email: false

This maps to `{ "notifications" => { "email" => false } }` in Ruby.

We currently set it to:

  notifications:
    email:
      - false

This maps to `{ "notifications" => { "email" => [false] } }` (note that `false`
is in an array). Because of this, Travis does not turn off email notifications.

Docs: http://docs.travis-ci.com/user/notifications/