Fix Circle deploys by removing redundant remote
Problem:

The circle configuration tries to deploy the build to the heroku staging
environment if it finishes successfully.

Because both the .circle.yml and bin/setup scripts
define the staging remote for git,
the build fails with a warning about the duplicate remote.

Solution:

Only set the staging remote once, in the setup script.
Remove the redundant source definition from `circle.yml`