Make new apps "deployable to Heroku" by default
This change makes new suspender apps "deployable to Heroku" by default,
moving the Heroku setup in the local machine to be run by default (i.e.
with the --heroku flag off). All the setup that requires to hit the
Heroku API to create the staging and production apps is kept to be run
only with the flag on.

Also in this commit, the specs moved from the "suspends a project for
Heroku" block in the heroku feature spec file are split into their own
specs in the new_project feature spec file to make them more
descriptive. Some of the names of the methods were modified to be more
descriptive and consistent with the result they are providing.