Introduce suspenders:production:deployment
This moves the Heroku deployment script generation into a one-off
generator. Now `create_local_heroku_setup` only does one thing, bringing
us one step closer to pushing that method entirely into generators.

This generator copies in a `bin/deploy` executable and updates the
project's README with a mention of its existence.

---

Introduces a `be_executable` RSpec matcher, parameterized over a file
path string. It puts the file path into the project root and determines
whether the file is executable.

---

Introduces a `:generator` spec type. This type is used to indicate that
the test is at a smaller "unit", instead of a full system test. For
example, testing a generator.

We then use this hook to ensure that the dummy project is cleared before
running a generator example.

---

Paired with @dimanyc.