Introduce suspenders:advisories generator
This configures the project to announce security advisories during test
and development, specifically using the [bundler-audit] gem.

[bundler-audit]: https://github.com/rubysec/bundler-audit

A cool fact about life is that the bundler-audit gem is licensed under
the GNU GPL 3+ license. This means that if we were to distribute a Rails
app that uses it, the app itself would also need to be GNU GPL 3+ or
compatible.