Stop customizing error pages
This did three things:

- Specify the charset in the HTML body.
- Specify the viewport.
- Remove comments.

The charset was removed from Rails in
https://github.com/rails/rails/commit/38f669766cb4d328e121afab020f4a52ca45421f .
It's not needed in the HTML if it's sent from the server. It's best to
send it from the server, in fact; otherwise the browser will parse the
first _n_ bytes of the file in a default charset, hit the charset meta
tag, then re-parse from the top under the new charset. We can remove
this meta tag now.

The viewport is in the default error page HTML as of
https://github.com/rails/rails/commit/7166507eae3d12c22e04a99ddd70972419643cd3 .

That leaves us with the comments. It'd be nice if they weren't there. I
don't know if that's worth a generator.