Move heredoc into templates
This adds a `templates/partials` directory, for storing incomplete
templates. They are named after their generator; for example, the
`runner` generator has templates `runner_readme.md` and
`runner_setup.rb`.

Extract all the heredoc into these templates. This gives us an easier
view into the content and more access to syntax highlighting in more
places (e.g. GitHub).

Along the way, backfill tests for generators without any: `ci` and
`db_optimizations. This exposed an issue with the `ci` generator: it did
not work with the default Rails `test/test_helper.rb`, and it mixed
SimpleCov in various places. So this is a substantial change to that
one generator.