Describe generators
This adds a description to all of our generators, except for the main
one. The main one will continue to use the `USAGE` file.

The descriptions are stored in a Markdown file under
`templates/descriptions`, named after the generator. For example,
`Suspenders::ViewsGenerator` is described by
`templates/descriptions/views.md`.

A little metaprogramming is used in the parent class to make this
seamless for the generators.

---

While here, move the `remove_prior_config` method up, so we don't remove
the config that we had just added.