Convert remaining generator tests (#1075)
In this commit, we are converting the remaining feature tests disguised as
generator tests to actual generator tests that run reasonably fast!

All generator tests from now on should follow this pattern.

In CI, the test suite used to run in ~36 minutes. Now it runs in ~12
minutes, so it's about 3 times faster. That time should continue to
drop as we extract new generators from AppGenerator.