Revert spring stop command
In this commit, we've removed the spring stop command:

427228ad544a97f28244e2696174fa1b2ac9d803

I was wrong and this is still needed because the suspenders
command does not include the bundled environment of suspenders
itself. Therefore, there's still a chance spring (from Rails 6.0)
would be running and that would make the suspenders:views generator
hang forever, so the suspenders command would fail to finish
completely.